scripts: Add support for git-submodule package feeds
Adding git repositories as submodules instead of a normal clone is more
suitable in most cases. It allows easier version control of the used
feed revision within a (custom) firmware build if an upstream repository
of the feed exists and does not store a redundant copy in the super
repository in this case.

Signed-off-by: Linus Lüssing <linus.luessing@web.de>