[package] base-files: If proto dhcp call hotplug on ifup, not dhcp lease
Currently, the ifup-hotplug scripts are being called after obtaining a
dhcp lease in the case of an interface set to the protocol "dhcp". This
for one thing has the confusing (undesired?) result, that for instance
aliases for the same interface are only being added _after_ the router
got a dhcp lease from a dhcp server. Or in other words, if no dhcp
server is present, then no aliases will be set.

However, this might be one of the reasons of having assigned an alias,
to be able to let the router get an internet connection via dhcp but
also to be able to access the device if no dhcp server is present on the
same interface.

This commit moves the hotplug script call from the script called after
obtaining the dhcp lease to the network configuration script where the
interface was just set up and the dhcp client started.

Signed-off-by: Linus Lüssing <linus.luessing@web.de>