Filter various multicast types from clients into the mesh
With 130 nodes and 400 clients, flooded multicast packets are starting to give
us a hard time. Average rates on a bat0 interface of a gateway are:
10972.16 bits/sec, 11.30 packets/sec

Which together with various extra headers (e.g. batman-adv + fastd VPN or wifi
headers) are becoming quite a burden on the limited 802.11 wifi broadcast
spectrum and common DSL uplinks.

Therefore we are filtering all multicast (including broadcast packets), execept
the ones relevant for routing:

* ARP, DHCPv4/v6, IGMP, ICMP(v6), OSPF, BABEL and Bittorrent Local Peer
    Discovery